Fixed pagination helper active state

+++ helpers/pagination.php
@@ -58,7 +58,7 @@ class Pagination extends Helper {
 
         // Prevent dividing if there are zero items.
         if ( $per_page > 0 ) {
-            $page_count = ceil( $items / $per_page );
+            $page_count = (int) ceil( $items / $per_page );
         }
         else {
             $page_count = 1;
